Add admin check

+++ routes.py
@@ -91,62 +91,68 @@ def register():
 
 @app.route("/add_course", methods=["POST"])
 def add_course():
-    course_name = request.form["course_name"]
-    if courses.add_course(course_name):
-        return redirect("/")
+    if users.is_admin():
+        course_name = request.form["course_name"]
+        if courses.add_course(course_name):
+            return redirect("/")
     return render_template("/", message="Kurssin lisääminen ei onnistu")
+
 
 @app.route("/delete_course")
 def delete_course():
-    if courses.delete_course(session["course_id"]):
-        return redirect("/")
+    if users.is_admin():
+        if courses.delete_course(session["course_id"]):
+            return redirect("/")
     return render_template("/", message="Kurssin poistaminen ei onnistu")
 
 @app.route("/add_textmaterial", methods=["POST"])
 def add_textmaterial():
-    print("asdfasdf")
-    textmaterial = request.form["textmaterial"]
-    course_id = session["course_id"]
-    if courses.add_textmaterial(course_id, textmaterial):
-        return redirect(f"/courses/{course_id}")
+    if users.is_admin():
+        textmaterial = request.form["textmaterial"]
+        course_id = session["course_id"]
+        if courses.add_textmaterial(course_id, textmaterial):
+            return redirect(f"/courses/{course_id}")
     return render_template("error.html", message="Materiaalin lisääminen ei onnistu")
 
 @app.route("/add_textquestion", methods=["POST"])
 def add_textquestion():
-    textquestion = request.form["textquestion"]
-    textanswer = request.form["textanswer"]
-    course_id = session["course_id"]
-    if questions.add_textquestion(course_id, textquestion, textanswer):
-        return redirect(f"/courses/{course_id}")
+    if users.is_admin():
+        textquestion = request.form["textquestion"]
+        textanswer = request.form["textanswer"]
+        course_id = session["course_id"]
+        if questions.add_textquestion(course_id, textquestion, textanswer):
+            return redirect(f"/courses/{course_id}")
     return render_template("error.html", message="Kysymyksen lisääminen ei onnistu")
 
 @app.route("/delete_textquestion", methods=["POST"])
 def delete_textquestion():
-    course_id = session["course_id"]
-    question_id = request.form["question_id"]
-    if questions.delete_textquestion(question_id):
-        return redirect(f"/courses/{course_id}")
+    if users.is_admin():
+        course_id = session["course_id"]
+        question_id = request.form["question_id"]
+        if questions.delete_textquestion(question_id):
+            return redirect(f"/courses/{course_id}")
     return render_template("error.html", message="Kysymyksen poistaminen ei onnistu")
 
 @app.route("/add_multiquestion", methods=["POST"])
 def add_multiquestion():
-    print("asdfasd")
-    multiquestion = request.form["multiquestion"]
-    choice1 = request.form["choice1"]
-    choice2 = request.form["choice2"]
-    choice3 = request.form["choice3"]
-    multianswer = request.form["multianswer"]
-    course_id = session["course_id"]
-    if questions.add_multiquestion(course_id, multiquestion, choice1, choice2, choice3, multianswer):
-        return redirect(f"/courses/{course_id}")
+    if users.is_admin():
+        multiquestion = request.form["multiquestion"]
+        choice1 = request.form["choice1"]
+        choice2 = request.form["choice2"]
+        choice3 = request.form["choice3"]
+        multianswer = request.form["multianswer"]
+        course_id = session["course_id"]
+        if questions.add_multiquestion(course_id, multiquestion, choice1, choice2, choice3, multianswer):
+            return redirect(f"/courses/{course_id}")
     return render_template("error.html", message="Kysymyksen lisääminen ei onnistu")
 
 @app.route("/delete_multiquestion", methods=["POST"])
 def delete_multiquestion():
-    course_id = session["course_id"]
-    question_id = request.form["question_id"]
-    if questions.delete_multiquestion(question_id):
-        return redirect(f"/courses/{course_id}")
+    if users.is_admin():
+        course_id = session["course_id"]
+        question_id = request.form["question_id"]
+        if questions.delete_multiquestion(question_id):
+            return redirect(f"/courses/{course_id}")
     return render_template("error.html", message="Kysymyksen poistaminen ei onnistu")
 
 @app.route("/answer_multiquestion", methods=["POST"])
